Accuracy fix for single Chunked Prefill Sequence

+++ tests/basic_correctness/test_chunked_prefill.py
@@ -24,7 +24,7 @@ MODELS = [
 
 
 @pytest.mark.parametrize("model", MODELS)
-@pytest.mark.parametrize("dtype", ["bfloat16"])
+@pytest.mark.parametrize("dtype", ["float"])
 @pytest.mark.parametrize("max_tokens", [32])
 @pytest.mark.parametrize("chunked_prefill_token_size", [4,])
 @pytest.mark.parametrize("enforce_eager", [True])
